Typing fixes for CI's mypy; make check runs mypy too
posterior.reliability returns a float, the per-axis count is an int, and the marginal
loop variable no longer shadows a float. make check now runs mypy as CI does (1.3 s
warm, 56 s cold), so a green make check means a green CI typecheck.

+++ src/life_agent/core/posterior.py
@@ -40,8 +40,8 @@ def temper_scales(groups: Sequence[int]) -> list[float]:
 
 def reliability(rho: float, obs: Mapping[str, Any]) -> float:
     """The chance this observation names the truth."""
-    return (rho * obs["authority"] * obs["subject_factor"] * obs["time_factor"]
-            * obs.get("competition_factor", 1.0))
+    return float(rho * obs["authority"] * obs["subject_factor"] * obs["time_factor"]
+                 * obs.get("competition_factor", 1.0))
 
 
 def log_match_miss(r: float, scale: float) -> tuple[float, float]:
